pyproject.toml
[tool.poetry]
name = "ad-diwan"
version = "0.1.0"
description = "Arabic poetry"
authors = ["Yousef Abu Shanab <josephyousef249@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "ad_diwan" }]
homepage = "https://github.com/al-arabiyya/ad-diwan"
repository = "https://github.com/al-arabiyya/ad-diwan"
documentation = "https://github.com/al-arabiyya/ad-diwan"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5.2.1"
django-filter = ">=25.1"
djangorestframework = ">=3.16.0"
drf-restwind = ">=0.2.0"
wagtail = ">=7.0"
wagtail-tw-blocks = ">=0.2.5"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = ">=6.0.0"
black = ">=25.1.0"
ruff = ">=0.11.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
